Modified GUI
Updated the GUI to remove the initial file prompt when starting the
application.  A new XML file can now be loaded whenever desired.  Also
made the tree view include both the elements and attributes.  This
simplified the layout a bit and makes it easier to add features down the
road.