Removed requirement to generate an additional class.
The program will now generate the XML file parsing in the root node class instead of creating the separate main class. This simplifies the code a bit. Note: The GUI still needs to be updated to remove the Main class name text box. Also added a bunch of overloads so that multiple options can be used to read/write XMl.