Bug Fixes and added support for namespaces.
Fixed a bug where Text inside elements was not being see correctly. The ElementInfo now stores the full name but also the namespace and local name separate.  Also added a button so that you can view the unique values contained in all the nodes. Also updated the gui with a TabControl so that you will have a separate tab for each namespace.